merge: fold zed_depth_map into zed_depth_costmap, rename to zed_stereo_nav
